requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
diff-match-patch==20121119
Django==2.0.1
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.4.9
Pillow==5.0.0
pycodestyle==2.4.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2017.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
